+++ Web/client/src/components/css/student.css
@@ -10,6 +10,4 @@
     display: flex;
 }
 
-.userInfo{
-    margin-top: 5%;
-}
+
